requirements.txt
ansible==2.10.7
ansible-base==2.10.10
asn1crypto==1.4.0
cffi==1.14.5
cryptography==2.7
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.19.5
packaging==20.9
pandas==0.24.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
